+++ react/components/BuyButton/Wrapper.js
@@ -1,14 +1,47 @@
 import React, { useContext } from 'react'
 import { ProductContext } from 'vtex.product-context'
-import { path, isEmpty, compose } from 'ramda'
+import { path, isEmpty, compose, pathOr } from 'ramda'
 import { FormattedMessage, injectIntl } from 'react-intl'
 import { withToast } from 'vtex.styleguide'
 import { orderFormConsumer } from 'vtex.store-resources/OrderFormContext'
-
-import gql from 'graphql-tag'
+import ProductPrice from '../ProductPrice'
 import { graphql } from 'react-apollo'
 
 import { BuyButton } from './index'
+import { transformAssemblyOptions } from './assemblyUtils'
+import addToCartMutation from './mutations/addToCart.gql'
+import setOpenMinicartMutation from './mutations/setOpenMinicart.gql'
+
+const BuyButtonMessage = ({ showItemsPrice, skuItems }) => {
+  if (!showItemsPrice) {
+    return <FormattedMessage id="store/buy-button.add-to-cart" />
+  }
+
+  const totalPrice = skuItems.reduce((acc, item) => {
+    const itemPrice = item.price * item.quantity
+    const addedAssemblyOptions = pathOr([], ['assemblyOptions', 'added'], item)
+    return (
+      acc +
+      addedAssemblyOptions.reduce(
+        (childAcc, option) =>
+          childAcc +
+          option.item.sellingPrice * option.normalizedQuantity * item.quantity,
+        itemPrice
+      )
+    )
+  }, 0)
+
+  return (
+    <div className="flex w-100 justify-between items-center">
+      <FormattedMessage id="store/buy-button.add-to-cart" />
+      <ProductPrice
+        showLabels={false}
+        showListPrice={false}
+        sellingPrice={totalPrice}
+      />
+    </div>
+  )
+}
 
 const BuyButtonWrapper = props => {
   const {
@@ -28,7 +61,12 @@ const BuyButtonWrapper = props => {
       return props
     }
 
-    const { product, selectedItem, selectedQuantity } = valuesFromContext
+    const {
+      product,
+      selectedItem,
+      selectedQuantity,
+      assemblyOptions,
+    } = valuesFromContext
 
     const sellerId = path(['sellers', 0, 'sellerId'], selectedItem)
     const commertialOffer = path(
@@ -58,6 +96,7 @@ const BuyButtonWrapper = props => {
             ['sellers', '0', 'commertialOffer', 'ListPrice'],
             selectedItem
           ),
+          ...transformAssemblyOptions(assemblyOptions, commertialOffer.Price),
         },
       ]
 
@@ -68,7 +107,7 @@ const BuyButtonWrapper = props => {
       available: props.available == null ? available : props.available,
     }
   }
-
+  const buttonProps = buyButtonProps()
   return (
     <BuyButton
       intl={intl}
@@ -77,47 +116,33 @@ const BuyButtonWrapper = props => {
       onAddFinish={onAddFinish}
       showToast={showToast}
       orderFormContext={orderFormContext}
-      {...buyButtonProps()}
+      {...buttonProps}
     >
-      {children || <FormattedMessage id="store/buy-button.add-to-cart" />}
+      {children || (
+        <BuyButtonMessage
+          showItemsPrice={props.showItemsPrice}
+          skuItems={buttonProps.skuItems}
+        />
+      )}
     </BuyButton>
   )
 }
 
-export const ADD_TO_CART_MUTATION = gql`
-  mutation addToCart($items: [MinicartItem]) {
-    addToCart(items: $items) @client
-  }
-`
-
-export const OPEN_CART_MUTATION = gql`
-  mutation setMinicartOpen($isOpen: Boolean!) {
-    setMinicartOpen(isOpen: $isOpen) @client
-  }
-`
-
-const withAddToCart = graphql(ADD_TO_CART_MUTATION, {
+const withAddToCart = graphql(addToCartMutation, {
   name: 'addToCart',
   props: ({ addToCart }) => ({
     addToCart: items => addToCart({ variables: { items } }),
   }),
 })
 
-const withOpenMinicart = graphql(OPEN_CART_MUTATION, {
+const withOpenMinicart = graphql(setOpenMinicartMutation, {
   name: 'setMinicartOpen',
   props: ({ setMinicartOpen }) => ({
     setMinicartOpen: isOpen => setMinicartOpen({ variables: { isOpen } }),
   }),
 })
 
-const withMutation = graphql(ADD_TO_CART_MUTATION, {
-  props: ({ mutate }) => ({
-    addToCart: items => mutate({ variables: { items } }),
-  }),
-})
-
 export default compose(
-  withMutation,
   withAddToCart,
   withOpenMinicart,
   withToast,